+++ R/types.R
@@ -179,6 +179,11 @@ type_object <- function(
 convert_from_type <- function(x, type) {
   if (S7_inherits(type, TypeArray)) {
     if (S7_inherits(type@items, TypeBasic)) {
+      if (!type@items@required) {
+        is_null <- map_lgl(x, is.null)
+        x[is_null] <- list(NA)
+      }
+
       switch(
         type@items@type,
         boolean = as.logical(x),